requirements_old - Copy.txt
fastapi==0.115.6
uvicorn==0.32.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
twilio==9.3.7
pydantic==2.10.3
pydantic-settings==2.6.1

